+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+- normalize error logger messages, use `inspect` for now
+
 ## [0.1.0] - 2016-11-15
 ### Added
 - base json backend implementation